+++ Source/Bake/Cooking/Cooks/GitHub/GitHubReleaseCook.cs
@@ -114,6 +114,22 @@ namespace Bake.Cooking.Cooks.GitHub
                     await file.GetHashAsync(HashAlgorithm.SHA256, cancellationToken)));
             }
 
+            var containerArtifacts = recipe.Artifacts
+                .OfType<ContainerArtifact>()
+                .ToArray();
+            if (containerArtifacts.Any())
+            {
+                stringBuilder.AppendLine("### Containers");
+                foreach (var containerArtifact in containerArtifacts)
+                {
+                    stringBuilder.AppendLine($"* `{containerArtifact.Key.Name}`");
+                    foreach (var tag in containerArtifact.Tags)
+                    {
+                        stringBuilder.AppendLine($"  * `{tag}`");
+                    }
+                }
+            }
+
             if (releaseFiles.Any())
             {
                 stringBuilder.AppendLine("### Files");